Добавит асинхронные action. Подключит Redux Thunk и редьюсеры
